feat: Release v1.0.0

+++ .github/set-version.sh
@@ -3,7 +3,7 @@
 SCRIPT_BASE="$(cd "$( dirname "$0")" && pwd )"
 ROOT=${SCRIPT_BASE}/..
 
-# Exit immediatly if any command exits with a non-zero status
+# Exit immediately if any command exits with a non-zero status
 set -e
 
 # Usage
@@ -15,29 +15,32 @@ print_usage() {
     echo ""
 }
 
-# if less than one arguments supplied, display usage
+# if less than one argument supplied, display usage
 if [  $# -lt 1 ]
 then
     print_usage
     exit 1
 fi
 
-# check whether user had supplied -h or --help . If yes display usage
+# check whether user had supplied -h or --help. If yes, display usage
 if [ "$1" == "--help" ] || [ "$1" == "-h" ]; then
     print_usage
     exit 0
 fi
 
-# NEW_VERSION=$(echo "$1" | sed -e 's/-beta\./.b/' | sed -e 's/-alpha\./.a/')
+echo "Setting new version to $1"
 
 # Set version in galaxy.yml
-grep -E '^version: (.+)$' "$ROOT/galaxy.yml" >/dev/null
+echo "Updating galaxy.yml"
+grep -E '^version: (.+)$' "$ROOT/galaxy.yml" || exit 1
 sed -i.bak -E "s/^version: (.+)$/version: $1/" "$ROOT/galaxy.yml" && rm "$ROOT/galaxy.yml.bak"
 
 # Set version in docs/source/index.rst
-grep -E '^Version: (.+)$' "$ROOT/docs/source/index.rst" > /dev/null
+echo "Updating docs/source/index.rst"
+grep -E '^Version: (.+)$' "$ROOT/docs/source/index.rst" || exit 1
 sed -i.bak -E "s/^Version: (.+)$/Version: $1/" "$ROOT/docs/source/index.rst" && rm "$ROOT/docs/source/index.rst.bak"
 
 # Set version in pyproject.toml
-grep -E '^version = ".+"$' "$ROOT/pyproject.toml" >/dev/null
+echo "Updating pyproject.toml"
+grep -E '^version = ".+"$' "$ROOT/pyproject.toml" || exit 1
 sed -i.bak -E "s/^version = \".+\"$/version = \"$1\"/" "$ROOT/pyproject.toml" && rm "$ROOT/pyproject.toml.bak"
